feat: 모델 사진 확대 dialog 컴포넌트 분리 및 전역 상태 연동
