+++ src/Widgets/FittingModel/Ui/ModelZoomDialog/ModelZoomDialog.tsx
@@ -1,0 +1,19 @@
+import { Dialog } from '@/Shared/Components';
+
+import { ModelZoomDialogContent } from './ModelZoomDialog.content';
+import { ModelZoomDialogTrigger } from './ModelZoomDialog.trigger';
+
+type ModelZoomDialogProps = {
+  dialogTriggerClassName?: string;
+};
+
+export const ModelZoomDialog = ({
+  dialogTriggerClassName,
+}: ModelZoomDialogProps) => {
+  return (
+    <Dialog>
+      <ModelZoomDialogTrigger dialogTriggerClassName={dialogTriggerClassName} />
+      <ModelZoomDialogContent />
+    </Dialog>
+  );
+};
